chore: moved impiccato_definitivo.py into main.py
